fix: unbreak CI (Lint & Type Check jobs have been failing for 17+ days)
- writers.py: drop redundant int() around round() (round() already
  returns int on the single-arg form) -- ruff RUF046.
- test_table_converter.py: prefix the intentionally-unused unpacked
  force column with _ -- ruff RUF059.
- pyproject.toml: bump mypy's python_version 3.10 -> 3.12. numpy's
  bundled stubs now use PEP 695 `type` statements (3.12+ syntax); mypy
  parses third-party stubs under the same configured target version,
  so `import numpy` failed regardless of this project's own 3.10
  minimum (requires-python).

None of this was introduced by the PET merge -- confirmed via
`gh run list` that both CI jobs were already red on main 17 days ago
(run 29069060529), just never fixed.

+++ python/src/backmap_prep/writers.py
@@ -290,14 +290,14 @@ def _format_dihedral_coeff(dihtype: DihedralTypeInfo, dihedral_styles: list[str]
         return f"dihedral_coeff {dihtype.type_id} {dihtype.keyword} {coeffs}\n"
     if dihtype.style == "charmm":
         k_val, n_val, delta = dihtype.params[:3]
-        shift = int(round(delta))
+        shift = round(delta)
         coeffs = f"{k_val:.6f} {int(n_val)} {shift} 1.0"
         if hybrid:
             return f"dihedral_coeff {dihtype.type_id} charmm {coeffs}\n"
         return f"dihedral_coeff {dihtype.type_id} {coeffs}\n"
     if dihtype.style == "backmap/charmm":
         k_val, n_val, delta = dihtype.params[:3]
-        shift = int(round(delta))
+        shift = round(delta)
         coeffs = f"{k_val:.6f} {int(n_val)} {shift} 1.0"
         if hybrid:
             return f"dihedral_coeff {dihtype.type_id} backmap/charmm {dihtype.keyword} {coeffs}\n"
